AlgebraicPrelude

Description

This module provides drop-in replacement for Prelude module in base package, based on algebraic hierarchy provided by algebra package. You can use this module with NoImplicitPrelude language option.

This module implicitly exports following modules:

Numeric.Algebra module, except

fromInteger: this module exports Prelude's fromInteger to make number literals work properly. For fromInteger from algebra package, use fromInteger'.

(^) is renamed to (^^), and (^) is redefined as pow.

The module Numeric.Algebra.Unital.UnitNormalForm, except for normalize; hence its name is too general, we export it as normaliseUnit.

Following modules are exported as-is:

Numeric.Decidable.Associates

Numeric.Decidable.Units

Numeric.Decidable.Zero

Numeric.Domain.Class

Numeric.Domain.Euclidean

Numeric.Domain.GCD

Numeric.Domain.Integral

Numeric.Domain.PID

Numeric.Domain.UFD

Numeric.Field.Fraction

Numeric.Semiring.ZeroProduct

Non-numeric part of this module is almost same as BasicPrelude. But the following combinators are not generalized from Prelude:

String-specific functions: getArgs, getContents, getLine, interact, putStr, putStrLn, read, readFile, writeFile, lines, unlines, words and unwords.

(.) is just a function composition; not a Categorical composition.

Basic types and renamed operations

type Rational = Fraction Integer

We use Fraction instead of Ratio for consistency.

fromInteger :: Num r => Integer -> r

To work with Num literals.

fromInteger' :: Ring r => Integer -> r

algebra package's original fromInteger.

fromRational :: DivisionRing r => Rational -> r

normaliseUnit :: UnitNormalForm r => r -> r

(^) :: Unital r => r -> Natural -> r infixr 8

Specialised version of pow which takes Naturals as a power.

(^^) :: Division r => r -> Integer -> r infixr 8

The original power function (^) of algebra

Combinator to use with RebindableSyntax extensions.

ifThenElse :: Bool -> a -> a -> a

Wrapper types for conversion between Num family

newtype WrapNum a

Wrapping Prelude's numerical types to treat with Algebra hierachy.

For Field or Euclidean instances, see WrapIntegral and WrapField.

N.B. This type provides a mean to convert from Nums to Rings, but there is no guarantee that WrapNum a is actually ring. For example, due to precision limitation, WrapPreldue Double even fails to be semigroup! For another simpler example, even though Natural comes with Num instance, but it doesn't support negate, so it cannot be Group.

Constructors

WrapNum

Fields

unwrapNum :: a

newtype WrapFractional a

Similar to WrapNum, but produces Field instances from Fractionals.

See also: WrapIntegral and WrapNum.

Constructors

WrapFractional

Fields

unwrapFractional :: a

newtype WrapIntegral a

Similar to WrapNum, but produces Euclidean instances from Integrals.

See also: WrapFractional and WrapNum.

Constructors

WrapIntegral

Fields

unwrapIntegral :: a

newtype WrapAlgebra a

Turning types from Algebra into Prelude's Num instances.

N.B. Since Real's toRational constraint is too tight, we won't provide the inverse of WrapIntegral and provide Fractional instance only.

Constructors

WrapAlgebra

Fields

unwrapAlgebra :: a

newtype Add a

Monoid instances for Additives. N.B. Unlike WrapNum, Num instance is just inhereted from the unwrapped data.

Constructors

Add

Fields

runAdd :: a

newtype Mult a

Monoid instances for Additives. N.B. Unlike WrapNum, Num instance is just inhereted from the unwrapped data.

Constructors

Mult

Fields

runMult :: a

Old Prelude's Numeric type classes and functions, without confliction

class Num a where

Basic numeric class.

Minimal complete definition

(+), (*), abs, signum, fromInteger, (negate | (-))

Methods

abs :: a -> a

Absolute value.

signum :: a -> a

Sign of a number. The functions abs and signum should satisfy the law:

abs x * signum x == x

For real numbers, the signum is either -1 (negative), 0 (zero) or 1 (positive).

class (Real a, Enum a) => Integral a where

Integral numbers, supporting integer division.

Minimal complete definition

quotRem, toInteger

Methods

div :: a -> a -> a infixl 7

integer division truncated toward negative infinity

mod :: a -> a -> a infixl 7

integer modulus, satisfying

(x `div` y)*y + (x `mod` y) == x

toInteger :: a -> Integer

conversion to Integer

toInteger :: Integral a => a -> Integer

conversion to Integer

class (Num a, Ord a) => Real a where

Minimal complete definition

toRational

Methods

toRational :: a -> Rational

the rational equivalent of its real argument with full precision

class Num a => Fractional a

Fractional numbers, supporting real division.

Minimal complete definition

fromRational, (recip | (/))

class Fractional a => Floating a where

Trigonometric and hyperbolic functions and related functions.

Minimal complete definition

pi, exp, log, sin, cos, asin, acos, atan, sinh, cosh, asinh, acosh, atanh

Methods

pi :: a

exp :: a -> a

log :: a -> a

sqrt :: a -> a

(**) :: a -> a -> a infixr 8

logBase :: a -> a -> a

sin :: a -> a

cos :: a -> a

tan :: a -> a

asin :: a -> a

acos :: a -> a

atan :: a -> a

sinh :: a -> a

cosh :: a -> a

tanh :: a -> a

asinh :: a -> a

acosh :: a -> a

atanh :: a -> a

class (Real a, Fractional a) => RealFrac a where

Extracting components of fractions.

Minimal complete definition

properFraction

Methods

properFraction :: Integral b => a -> (b, a)

The function properFraction takes a real fractional number x and returns a pair (n,f) such that x = n+f, and:

n is an integral number with the same sign as x; and

f is a fraction with the same type and sign as x, and with absolute value less than 1.

The default definitions of the ceiling, floor, truncate and round functions are in terms of properFraction.

truncate :: Integral b => a -> b

truncate x returns the integer nearest x between zero and x

round :: Integral b => a -> b

round x returns the nearest integer to x; the even integer if x is equidistant between two integers

ceiling :: Integral b => a -> b

ceiling x returns the least integer not less than x

floor :: Integral b => a -> b

floor x returns the greatest integer not greater than x

class (RealFrac a, Floating a) => RealFloat a where

Efficient, machine-independent access to the components of a floating-point number.

Minimal complete definition

floatRadix, floatDigits, floatRange, decodeFloat, encodeFloat, isNaN, isInfinite, isDenormalized, isNegativeZero, isIEEE

Methods

floatRadix :: a -> Integer

a constant function, returning the radix of the representation (often 2)

floatDigits :: a -> Int

a constant function, returning the number of digits of floatRadix in the significand

floatRange :: a -> (Int, Int)

a constant function, returning the lowest and highest values the exponent may assume

decodeFloat :: a -> (Integer, Int)

The function decodeFloat applied to a real floating-point number returns the significand expressed as an Integer and an appropriately scaled exponent (an Int). If decodeFloat x yields (m,n), then x is equal in value to m*b^^n, where b is the floating-point radix, and furthermore, either m and n are both zero or else b^(d-1) <= abs m < b^d, where d is the value of floatDigits x. In particular, decodeFloat 0 = (0,0). If the type contains a negative zero, also decodeFloat (-0.0) = (0,0). The result of decodeFloat x is unspecified if either of isNaN x or isInfinite x is True.

encodeFloat :: Integer -> Int -> a

encodeFloat performs the inverse of decodeFloat in the sense that for finite x with the exception of -0.0, uncurry encodeFloat (decodeFloat x) = x. encodeFloat m n is one of the two closest representable floating-point numbers to m*b^^n (or ±Infinity if overflow occurs); usually the closer, but if m contains too many bits, the result may be rounded in the wrong direction.

exponent :: a -> Int

exponent corresponds to the second component of decodeFloat. exponent 0 = 0 and for finite nonzero x, exponent x = snd (decodeFloat x) + floatDigits x. If x is a finite floating-point number, it is equal in value to significand x * b ^^ exponent x, where b is the floating-point radix. The behaviour is unspecified on infinite or NaN values.

significand :: a -> a

The first component of decodeFloat, scaled to lie in the open interval (-1,1), either 0.0 or of absolute value >= 1/b, where b is the floating-point radix. The behaviour is unspecified on infinite or NaN values.

scaleFloat :: Int -> a -> a

multiplies a floating-point number by an integer power of the radix

isNaN :: a -> Bool

True if the argument is an IEEE "not-a-number" (NaN) value

isInfinite :: a -> Bool

True if the argument is an IEEE infinity or negative infinity

isDenormalized :: a -> Bool

True if the argument is too small to be represented in normalized format

isNegativeZero :: a -> Bool

True if the argument is an IEEE negative zero

isIEEE :: a -> Bool

True if the argument is an IEEE floating point number

atan2 :: a -> a -> a

a version of arctangent taking two real floating-point arguments. For real floating x and y, atan2 y x computes the angle (from the positive x-axis) of the vector from the origin to the point (x,y). atan2 y x returns a value in the range [-pi, pi]. It follows the Common Lisp semantics for the origin when signed zeroes are supported. atan2 y 1, with y in a type that is RealFloat, should return the same value as atan y. A default definition of atan2 is provided, but implementors can provide a more accurate implementation.

Orphan instances

Euclidean d => Fractional (Fraction d)

Methods

(/) :: Fraction d -> Fraction d -> Fraction d

recip :: Fraction d -> Fraction d

fromRational :: Rational -> Fraction d

Euclidean a => Num (Fraction a)

Methods

(+) :: Fraction a -> Fraction a -> Fraction a

(-) :: Fraction a -> Fraction a -> Fraction a

(*) :: Fraction a -> Fraction a -> Fraction a

negate :: Fraction a -> Fraction a

abs :: Fraction a -> Fraction a

signum :: Fraction a -> Fraction a

fromInteger :: Integer -> Fraction a