AlgebraicPrelude

Description

This module provides drop-in replacement for Prelude module in base package, based on algebraic hierarchy provided by algebra package. You can use this module with NoImplicitPrelude language option.

This module implicitly exports following modules:

Numeric.Algebra module, except

fromInteger: this module exports Prelude's fromInteger to make number literals work properly. For fromInteger from algebra package, use fromInteger'.

(^) is renamed to (^^), and (^) is redefined as pow.

The module Numeric.Algebra.Unital.UnitNormalForm, except for normalize; hence its name is too general, we export it as normaliseUnit.

Following modules are exported as-is:

Numeric.Decidable.Associates

Numeric.Decidable.Units

Numeric.Decidable.Zero

Numeric.Domain.Class

Numeric.Domain.Euclidean

Numeric.Domain.GCD

Numeric.Domain.Integral

Numeric.Domain.PID

Numeric.Domain.UFD

Numeric.Field.Fraction

Numeric.Semiring.ZeroProduct

Non-numeric part of this module is almost same as BasicPrelude. But the following combinators are not generalized from Prelude:

String-specific functions: getArgs, getContents, getLine, interact, putStr, putStrLn, read, readFile, writeFile, lines, unlines, words and unwords.

(.) is just a function composition; not a Categorical composition.

Basic types and renamed operations

type Rational = Fraction Integer

We use Fraction instead of Ratio for consistency.

fromInteger :: Num r => Integer -> r

To work with Num literals.

fromInteger' :: Ring r => Integer -> r

algebra package's original fromInteger.

fromRational :: DivisionRing r => Rational -> r

normaliseUnit :: UnitNormalForm r => r -> r

(^) :: Unital r => r -> Natural -> r infixr 8

Specialised version of pow which takes Naturals as a power.

(^^) :: Division r => r -> Integer -> r infixr 8

The original power function (^) of algebra

Combinator to use with RebindableSyntax extensions.

ifThenElse :: Bool -> a -> a -> a

Wrapper types for conversion between Num family

newtype WrapNum a

Wrapping Prelude's numerical types to treat with Algebra hierachy.

For Field or Euclidean instances, see WrapIntegral and WrapField.

N.B. This type provides a mean to convert from Nums to Rings, but there is no guarantee that WrapNum a is actually ring. For example, due to precision limitation, WrapPreldue Double even fails to be semigroup! For another simpler example, even though Natural comes with Num instance, but it doesn't support negate, so it cannot be Group.

Constructors

WrapNum

Fields

unwrapNum :: a

newtype WrapFractional a

Similar to WrapNum, but produces Field instances from Fractionals.

See also: WrapIntegral and WrapNum.

Constructors

WrapFractional

Fields

unwrapFractional :: a

newtype WrapIntegral a

Similar to WrapNum, but produces Euclidean instances from Integrals.

See also: WrapFractional and WrapNum.

Constructors

WrapIntegral

Fields

unwrapIntegral :: a

newtype WrapAlgebra a

Turning types from Algebra into Prelude's Num instances.

N.B. Since Real's toRational constraint is too tight, we won't provide the inverse of WrapIntegral and provide Fractional instance only.

Constructors

WrapAlgebra

Fields

unwrapAlgebra :: a

newtype Add a

Monoid instances for Additives. N.B. Unlike WrapNum, Num instance is just inhereted from the unwrapped data.

Constructors

Add

Fields

runAdd :: a

newtype Mult a

Monoid instances for Additives. N.B. Unlike WrapNum, Num instance is just inhereted from the unwrapped data.

Constructors

Mult

Fields

runMult :: a

Old Prelude's Numeric type classes and functions, without confliction

class Num a where

Basic numeric class.

Minimal complete definition

(+), (*), abs, signum, fromInteger, (negate | (-))

class (Real a, Enum a) => Integral a where

Integral numbers, supporting integer division.

Minimal complete definition

quotRem, toInteger

toInteger :: Integral a => a -> Integer

conversion to Integer

class (Num a, Ord a) => Real a where

class Num a => Fractional a

Fractional numbers, supporting real division.

Minimal complete definition

fromRational, (recip | (/))

class Fractional a => Floating a where

Trigonometric and hyperbolic functions and related functions.

Minimal complete definition

pi, exp, log, sin, cos, asin, acos, atan, sinh, cosh, asinh, acosh, atanh

class (Real a, Fractional a) => RealFrac a where

Extracting components of fractions.

Minimal complete definition

properFraction

class (RealFrac a, Floating a) => RealFloat a where

Efficient, machine-independent access to the components of a floating-point number.

Minimal complete definition

floatRadix, floatDigits, floatRange, decodeFloat, encodeFloat, isNaN, isInfinite, isDenormalized, isNegativeZero, isIEEE

Orphan instances

Euclidean d => Fractional (Fraction d)

Methods

(/) :: Fraction d -> Fraction d -> Fraction d

recip :: Fraction d -> Fraction d

fromRational :: Rational -> Fraction d

Euclidean a => Num (Fraction a)

Methods

(+) :: Fraction a -> Fraction a -> Fraction a

(-) :: Fraction a -> Fraction a -> Fraction a

(*) :: Fraction a -> Fraction a -> Fraction a

negate :: Fraction a -> Fraction a

abs :: Fraction a -> Fraction a

signum :: Fraction a -> Fraction a

fromInteger :: Integer -> Fraction a