AlgebraicPrelude

Documentation

type Rational = FractionInteger

class Ord (Norm a) => Normed a where

Minimal complete definition

norm, liftNorm

Associated Types

type Norm a

Methods

norm :: a -> Norm a

liftNorm :: Norm a -> a

abs :: Normed a => a -> a

(^) :: Unital r => r -> Natural -> r infixr 8

Specialised version of pow which takes Naturals as a power.

(^^) :: Division r => r -> Integer -> r infixr 8

The original power function (^) of algebra

ifThenElse :: Bool -> a -> a -> a

fromInteger :: Num r => Integer -> r

To work with Num literals.

fromInteger' :: Ring r => Integer -> r

algebra package's original fromInteger.

fromRational :: DivisionRing r => Rational -> r

Old Prelude's Numeric type classes and functions, without confliction

class Num a

Basic numeric class.

Minimal complete definition

(+), (*), abs, signum, fromInteger, (negate | (-))

class (Real a, Enum a) => Integral a where

Integral numbers, supporting integer division.

Minimal complete definition

quotRem, toInteger

toInteger :: Integral a => a -> Integer

conversion to Integer

class Num a => Fractional a

Fractional numbers, supporting real division.

Minimal complete definition

fromRational, (recip | (/))

class Fractional a => Floating a where

Trigonometric and hyperbolic functions and related functions.

Minimal complete definition

pi, exp, log, sin, cos, asin, acos, atan, sinh, cosh, asinh, acosh, atanh

class (Real a, Fractional a) => RealFrac a where

Extracting components of fractions.

Minimal complete definition

properFraction

class (RealFrac a, Floating a) => RealFloat a where

Efficient, machine-independent access to the components of a floating-point number.

Minimal complete definition

floatRadix, floatDigits, floatRange, decodeFloat, encodeFloat, isNaN, isInfinite, isDenormalized, isNegativeZero, isIEEE

realToFrac :: (Real a, Fractional b) => a -> b

general coercion to fractional types